requirements.txt
_tflow_select==2.3.0
absl-py==0.9.0
appnope==0.1.2
argon2-cffi==20.1.0
astor==0.8.1
astroid==2.4.2
async_generator==1.10
attrs==20.1.0
backcall==0.2.0
blas==1.0
bleach==3.2.1
bzip2==1.0.8
c-ares==1.15.0
ca-certificates==2021.1.19
cairo==1.14.12
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cloudpickle==1.3.0
cycler==0.10.0
dbus==1.13.18
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
donkeycar==3.1.5
entrypoints==0.3
expat==2.2.10
ffmpeg==4.0
fid-score==0.1.3
fontconfig==2.13.0
freetype==2.10.2
future==0.18.2
gast==0.4.0
gettext==0.19.8.1
glib==2.65.0
graphite2==1.3.14
grpcio==1.31.0
gym-donkeycar==1.0.16
gym==0.17.2
h5py==2.8.0
harfbuzz==1.8.8
hdf5==1.10.2
icu==58.2
idna==2.10
imageio-ffmpeg==0.4.2
imageio==2.9.0
importlib-metadata==1.7.0
importlib_metadata==1.7.0
iniconfig==1.0.1
intel-openmp==2019.4
ipykernel==5.3.4
ipython==7.19.0
ipython_genutils==0.2.0
ipywidgets==7.5.1
isort==5.4.2
jasper==2.0.14
jedi==0.17.2
jinja2==2.11.2
jpeg==9b
jsonschema==3.2.0
jupyter==1.0.0
jupyter_client==6.1.7
jupyter_console==6.2.0
jupyter_core==4.7.0
jupyterlab_pygments==0.1.2
keras-applications==1.0.8
keras-base==2.3.1
keras-preprocessing==1.1.0
keras-vis==0.5.0
keras==2.3.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lcms2==2.11
libcxx==10.0.0
libedit==3.1.20191231
libffi==3.3
libgfortran==3.0.1
libiconv==1.16
libopenblas==0.3.10
libopencv==3.4.2
libopus==1.3.1
libpng==1.6.37
libprotobuf==3.12.4
libsodium==1.0.18
libtiff==4.1.0
libvpx==1.7.0
libxml2==2.9.10
lz4-c==1.9.2
markdown==3.2.2
markupsafe==1.1.1
matplotlib-base==3.3.1
matplotlib==3.3.1
mccabe==0.6.1
mistune==0.8.4
mkl-service==2.3.0
mkl==2019.4
mock==4.0.2
more-itertools==8.5.0
moviepy==1.0.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
ncurses==6.2
nest-asyncio==1.4.3
networkx==2.5
nomkl==3.0
notebook==6.1.5
numpy-base==1.19.1
numpy==1.19.1
olefile==0.46
opencv==3.4.2
openssl==1.1.1j
packaging==20.4
paho-mqtt==1.5.0
pandas==1.1.1
pandoc==2.11
pandocfilters==1.4.3
parso==0.7.0
pcre==8.44
pexpect==4.8.0
pickleshare==0.7.5
pillow==7.2.0
pip==20.2.2
pixman==0.40.0
pluggy==0.13.1
prettytable==0.7.2
proglog==0.1.9
prometheus_client==0.9.0
prompt-toolkit==3.0.8
prompt_toolkit==3.0.8
protobuf==3.12.4
ptyprocess==0.6.0
py-opencv==3.4.2
py==1.9.0
pycparser==2.20
pyglet==1.5.0
pygments==2.7.3
pylint==2.6.0
pyparsing==2.4.7
pyqt==5.9.2
pyrsistent==0.17.3
pytest-mock==3.3.1
pytest==6.0.1
python-dateutil==2.8.1
python==3.7.9
pytz==2020.1
pywavelets==1.1.1
pyyaml==5.3.1
pyzmq==20.0.0
qt==5.9.7
qtconsole==4.7.7
qtpy==1.9.0
readline==8.0
requests==2.24.0
scikit-image==0.17.2
scipy==1.3.2
send2trash==1.5.0
setuptools==49.6.0
simple-pid==0.2.4
sip==4.19.8
six==1.15.0
sqlite==3.33.0
tensorboard==1.13.1
tensorflow-base==1.13.1
tensorflow-estimator==1.13.0
tensorflow==1.13.1
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tifffile==2020.8.25
tk==8.6.10
toml==0.10.1
torch==1.8.1
torchvision==0.9.1
tornado==6.0.4
tqdm==4.48.2
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.10.0.0
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==1.0.1
wheel==0.35.1
widgetsnbextension==3.5.1
wrapt==1.11.2
xz==5.2.5
yaml==0.2.5
zeromq==4.3.3
zipp==3.1.0
zlib==1.2.11
zstd==1.4.5
